refactor(conversations): delete the pages/Conversations shim (Phase 6, partial)
The panel has lived at `features/conversations/Conversations.tsx` since Phase 1b;
remove the temporary `pages/Conversations.tsx` re-export shim and point every
consumer at the real module: `HumanPage`, `Accounts` (+ its `ConversationsPage`
alias), and the co-located Conversations/Accounts test suites and mocks.

Scope note: Phase 6's dead-state removal (`parallelStreamsByThread`,
`preserveLiveSubagentProse`, provider refs) is coupled to Phase 3 and is tracked
in the follow-up issue, not this PR.

Typecheck + lint clean; 119 tests across the Conversations render suite,
Accounts, and HumanPage pass with the repointed imports.

Refs: docs/plans/conversations-timeline-refactor.md Phase 6
