+++ app/src/pages/__tests__/Conversations.test.tsx
@@ -4,7 +4,7 @@ import {
   formatThreadLoadError,
   isComposerInteractionBlocked,
   isImeCompositionKeyEvent,
-} from '../Conversations';
+} from '../../features/conversations/Conversations';
 
 describe('isComposerInteractionBlocked', () => {
   it('blocks composer interaction while the selected thread is actively running', () => {
